+++ cmd/cortex/main.go
@@ -7,6 +7,7 @@ import (
 	"math/rand"
 	"os"
 	"runtime"
+	"sync"
 	"time"
 
 	"github.com/go-kit/kit/log/level"
@@ -99,12 +100,26 @@ func main() {
 
 	level.Info(util.Logger).Log("msg", "Starting Cortex", "version", version.Info())
 
-	if err := t.Run(); err != nil {
-		level.Error(util.Logger).Log("msg", "error running Cortex", "err", err)
+	var wg sync.WaitGroup
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		if err := t.Run(); err != nil {
+			level.Error(util.Logger).Log("msg", "error running Cortex", "err", err)
+		}
+	}()
+
+	if cfg.Target.IsJob() {
+		err = t.Stop()
+	}
+
+	wg.Wait()
+
+	if !cfg.Target.IsJob() {
+		err = t.Stop()
 	}
 
 	runtime.KeepAlive(ballast)
-	err = t.Stop()
 	util.CheckFatal("initializing cortex", err)
 }