+++ pkg/cortex/cortex.go
@@ -24,6 +24,7 @@ import (
 	config_client "github.com/cortexproject/cortex/pkg/configs/client"
 	"github.com/cortexproject/cortex/pkg/configs/db"
 	"github.com/cortexproject/cortex/pkg/distributor"
+	"github.com/cortexproject/cortex/pkg/flusher"
 	"github.com/cortexproject/cortex/pkg/ingester"
 	"github.com/cortexproject/cortex/pkg/ingester/client"
 	"github.com/cortexproject/cortex/pkg/querier"
@@ -66,6 +67,7 @@ type Config struct {
 	Querier        querier.Config           `yaml:"querier,omitempty"`
 	IngesterClient client.Config            `yaml:"ingester_client,omitempty"`
 	Ingester       ingester.Config          `yaml:"ingester,omitempty"`
+	Flusher        flusher.Config           `yaml:"flusher,omitempty"`
 	Storage        storage.Config           `yaml:"storage,omitempty"`
 	ChunkStore     chunk.StoreConfig        `yaml:"chunk_store,omitempty"`
 	Schema         chunk.SchemaConfig       `yaml:"schema,omitempty" doc:"hidden"` // Doc generation tool doesn't support it because part of the SchemaConfig doesn't support CLI flags (needs manual documentation)
@@ -101,6 +103,7 @@ func (c *Config) RegisterFlags(f *flag.FlagSet) {
 	c.Querier.RegisterFlags(f)
 	c.IngesterClient.RegisterFlags(f)
 	c.Ingester.RegisterFlags(f)
+	c.Flusher.RegisterFlags(f)
 	c.Storage.RegisterFlags(f)
 	c.ChunkStore.RegisterFlags(f)
 	c.Schema.RegisterFlags(f)
@@ -164,6 +167,7 @@ type Cortex struct {
 	overrides     *validation.Overrides
 	distributor   *distributor.Distributor
 	ingester      *ingester.Ingester
+	flusher       *flusher.Flusher
 	store         chunk.Store
 	worker        frontend.Worker
 	frontend      *frontend.Frontend