+++ pkg/ingester/series.go
@@ -24,11 +24,11 @@ func init() {
 	prometheus.MustRegister(createdChunks)
 }
 
-type memorySeries struct {
+type MemorySeries struct {
 	metric labels.Labels
 
 	// Sorted by start time, overlapping chunk ranges are forbidden.
-	chunkDescs []*desc
+	chunkDescs []*Desc
 
 	// Whether the current head chunk has already been finished.  If true,
 	// the current head chunk must not be modified anymore.
@@ -43,8 +43,8 @@ type memorySeries struct {
 
 // newMemorySeries returns a pointer to a newly allocated memorySeries for the
 // given metric.
-func newMemorySeries(m labels.Labels) *memorySeries {
-	return &memorySeries{
+func newMemorySeries(m labels.Labels) *MemorySeries {
+	return &MemorySeries{
 		metric:   m,
 		lastTime: model.Earliest,
 	}
@@ -52,7 +52,7 @@ func newMemorySeries(m labels.Labels) *memorySeries {
 
 // add adds a sample pair to the series, possibly creating a new chunk.
 // The caller must have locked the fingerprint of the series.
-func (s *memorySeries) add(v model.SamplePair) error {
+func (s *MemorySeries) add(v model.SamplePair) error {
 	// If sender has repeated the same timestamp, check more closely and perhaps return error.
 	if v.Timestamp == s.lastTime {
 		// If we don't know what the last sample value is, silently discard.
@@ -126,7 +126,7 @@ func firstAndLastTimes(c encoding.Chunk) (model.Time, model.Time, error) {
 // closeHead marks the head chunk closed. The caller must have locked
 // the fingerprint of the memorySeries. This method will panic if this
 // series has no chunk descriptors.
-func (s *memorySeries) closeHead(reason flushReason) {
+func (s *MemorySeries) closeHead(reason flushReason) {
 	s.chunkDescs[0].flushReason = reason
 	s.headChunkClosed = true
 }
@@ -134,14 +134,14 @@ func (s *memorySeries) closeHead(reason flushReason) {
 // firstTime returns the earliest known time for the series. The caller must have
 // locked the fingerprint of the memorySeries. This method will panic if this
 // series has no chunk descriptors.
-func (s *memorySeries) firstTime() model.Time {
+func (s *MemorySeries) firstTime() model.Time {
 	return s.chunkDescs[0].FirstTime
 }
 
 // Returns time of oldest chunk in the series, that isn't flushed. If there are
 // no chunks, or all chunks are flushed, returns 0.
 // The caller must have locked the fingerprint of the memorySeries.
-func (s *memorySeries) firstUnflushedChunkTime() model.Time {
+func (s *MemorySeries) firstUnflushedChunkTime() model.Time {
 	for _, c := range s.chunkDescs {
 		if !c.flushed {
 			return c.FirstTime
@@ -154,11 +154,11 @@ func (s *memorySeries) firstUnflushedChunkTime() model.Time {
 // head returns a pointer to the head chunk descriptor. The caller must have
 // locked the fingerprint of the memorySeries. This method will panic if this
 // series has no chunk descriptors.
-func (s *memorySeries) head() *desc {
+func (s *MemorySeries) head() *Desc {
 	return s.chunkDescs[len(s.chunkDescs)-1]
 }
 
-func (s *memorySeries) samplesForRange(from, through model.Time) ([]model.SamplePair, error) {
+func (s *MemorySeries) samplesForRange(from, through model.Time) ([]model.SamplePair, error) {
 	// Find first chunk with start time after "from".
 	fromIdx := sort.Search(len(s.chunkDescs), func(i int) bool {
 		return s.chunkDescs[i].FirstTime.After(from)
@@ -199,7 +199,7 @@ func (s *memorySeries) samplesForRange(from, through model.Time) ([]model.Sample
 	return values, nil
 }
 
-func (s *memorySeries) setChunks(descs []*desc) error {
+func (s *MemorySeries) setChunks(descs []*Desc) error {
 	if len(s.chunkDescs) != 0 {
 		return fmt.Errorf("series already has chunks")
 	}
@@ -211,11 +211,19 @@ func (s *memorySeries) setChunks(descs []*desc) error {
 	return nil
 }
 
-func (s *memorySeries) isStale() bool {
+func (s *MemorySeries) GetChunks() []*Desc {
+	return s.chunkDescs
+}
+
+func (s *MemorySeries) Metric() labels.Labels {
+	return s.metric
+}
+
+func (s *MemorySeries) isStale() bool {
 	return s.lastSampleValueSet && value.IsStaleNaN(float64(s.lastSampleValue))
 }
 
-type desc struct {
+type Desc struct {
 	C           encoding.Chunk // nil if chunk is evicted.
 	FirstTime   model.Time     // Timestamp of first sample. Populated at creation. Immutable.
 	LastTime    model.Time     // Timestamp of last sample. Populated at creation & on append.
@@ -224,8 +232,8 @@ type desc struct {
 	flushed     bool           // set to true when flush succeeds
 }
 
-func newDesc(c encoding.Chunk, firstTime model.Time, lastTime model.Time) *desc {
-	return &desc{
+func newDesc(c encoding.Chunk, firstTime model.Time, lastTime model.Time) *Desc {
+	return &Desc{
 		C:          c,
 		FirstTime:  firstTime,
 		LastTime:   lastTime,
@@ -236,7 +244,7 @@ func newDesc(c encoding.Chunk, firstTime model.Time, lastTime model.Time) *desc 
 // Add adds a sample pair to the underlying chunk. For safe concurrent access,
 // The chunk must be pinned, and the caller must have locked the fingerprint of
 // the series.
-func (d *desc) add(s model.SamplePair) (encoding.Chunk, error) {
+func (d *Desc) add(s model.SamplePair) (encoding.Chunk, error) {
 	cs, err := d.C.Add(s)
 	if err != nil {
 		return nil, err
@@ -250,8 +258,8 @@ func (d *desc) add(s model.SamplePair) (encoding.Chunk, error) {
 	return cs, nil
 }
 
-func (d *desc) slice(start, end model.Time) *desc {
-	return &desc{
+func (d *Desc) slice(start, end model.Time) *Desc {
+	return &Desc{
 		C:         d.C.Slice(start, end),
 		FirstTime: start,
 		LastTime:  end,