+++ pkg/ingester/series_map.go
@@ -21,15 +21,15 @@ type seriesMap struct {
 
 type shard struct {
 	mtx sync.Mutex
-	m   map[model.Fingerprint]*memorySeries
+	m   map[model.Fingerprint]*MemorySeries
 	// Align this struct.
-	pad [cacheLineSize - unsafe.Sizeof(sync.Mutex{}) - unsafe.Sizeof(map[model.Fingerprint]*memorySeries{})]byte
+	_ [cacheLineSize - unsafe.Sizeof(sync.Mutex{}) - unsafe.Sizeof(map[model.Fingerprint]*MemorySeries{})]byte
 }
 
-// fingerprintSeriesPair pairs a fingerprint with a memorySeries pointer.
-type fingerprintSeriesPair struct {
-	fp     model.Fingerprint
-	series *memorySeries
+// FingerprintSeriesPair pairs a fingerprint with a memorySeries pointer.
+type FingerprintSeriesPair struct {
+	Fingerprint model.Fingerprint
+	Series      *MemorySeries
 }
 
 // newSeriesMap returns a newly allocated empty seriesMap. To create a seriesMap
@@ -37,7 +37,7 @@ type fingerprintSeriesPair struct {
 func newSeriesMap() *seriesMap {
 	shards := make([]shard, seriesMapShards)
 	for i := 0; i < seriesMapShards; i++ {
-		shards[i].m = map[model.Fingerprint]*memorySeries{}
+		shards[i].m = map[model.Fingerprint]*MemorySeries{}
 	}
 	return &seriesMap{
 		shards: shards,
@@ -46,7 +46,7 @@ func newSeriesMap() *seriesMap {
 
 // get returns a memorySeries for a fingerprint. Return values have the same
 // semantics as the native Go map.
-func (sm *seriesMap) get(fp model.Fingerprint) (*memorySeries, bool) {
+func (sm *seriesMap) get(fp model.Fingerprint) (*MemorySeries, bool) {
 	shard := &sm.shards[util.HashFP(fp)%seriesMapShards]
 	shard.mtx.Lock()
 	ms, ok := shard.m[fp]
@@ -55,7 +55,7 @@ func (sm *seriesMap) get(fp model.Fingerprint) (*memorySeries, bool) {
 }
 
 // put adds a mapping to the seriesMap.
-func (sm *seriesMap) put(fp model.Fingerprint, s *memorySeries) {
+func (sm *seriesMap) put(fp model.Fingerprint, s *MemorySeries) {
 	shard := &sm.shards[util.HashFP(fp)%seriesMapShards]
 	shard.mtx.Lock()
 	_, ok := shard.m[fp]
@@ -79,21 +79,21 @@ func (sm *seriesMap) del(fp model.Fingerprint) {
 	}
 }
 
-// iter returns a channel that produces all mappings in the seriesMap. The
+// Iter returns a channel that produces all mappings in the seriesMap. The
 // channel will be closed once all fingerprints have been received. Not
 // consuming all fingerprints from the channel will leak a goroutine. The
 // semantics of concurrent modification of seriesMap is the similar as the one
 // for iterating over a map with a 'range' clause. However, if the next element
 // in iteration order is removed after the current element has been received
 // from the channel, it will still be produced by the channel.
-func (sm *seriesMap) iter() <-chan fingerprintSeriesPair {
-	ch := make(chan fingerprintSeriesPair)
+func (sm *seriesMap) Iter() <-chan FingerprintSeriesPair {
+	ch := make(chan FingerprintSeriesPair)
 	go func() {
 		for i := range sm.shards {
 			sm.shards[i].mtx.Lock()
 			for fp, ms := range sm.shards[i].m {
 				sm.shards[i].mtx.Unlock()
-				ch <- fingerprintSeriesPair{fp, ms}
+				ch <- FingerprintSeriesPair{fp, ms}
 				sm.shards[i].mtx.Lock()
 			}
 			sm.shards[i].mtx.Unlock()