+++ pkg/ingester/user_state_test.go
@@ -53,7 +53,7 @@ func TestForSeriesMatchingBatching(t *testing.T) {
 
 				total, batch, batches := 0, 0, 0
 				err = instance.forSeriesMatching(ctx, tc.matchers,
-					func(_ context.Context, _ model.Fingerprint, s *memorySeries) error {
+					func(_ context.Context, _ model.Fingerprint, s *MemorySeries) error {
 						batch++
 						return nil
 					},