+++ CHANGELOG.md
@@ -8,7 +8,10 @@
   * `--store.min-chunk-age` has been removed
   * `--querier.query-store-after` has been added in it's place.
 * [FEATURE] Added user sub rings to distribute users to a subset of ingesters. #1947
-  * `--experimental.distributor.user-subring-size`
+* [FEATURE] Flusher target to flush the WAL.
+  * `-flusher.wal-dir` for the WAL directory to recover from.
+  * `-flusher.concurrent-flushes` for number of concurrent flushes.
+  * `-flusher.flush-op-timeout` is duration after which a flush should timeout.
 * [ENHANCEMENT] Experimental TSDB: Export TSDB Syncer metrics from Compactor component, they are prefixed with `cortex_compactor_`. #2023
 * [ENHANCEMENT] Experimental TSDB: Added dedicated flag `-experimental.tsdb.bucket-store.tenant-sync-concurrency` to configure the maximum number of concurrent tenants for which blocks are synched. #2026
 * [ENHANCEMENT] Experimental TSDB: Expose metrics for objstore operations (prefixed with `cortex_<component>_thanos_objstore_`, component being one of `ingester`, `querier` and `compactor`). #2027